requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Werkzeug==2.0.1
PyMySQL==1.0.2
SQLAlchemy==1.4.23
cryptography==3.4.7
python-dotenv==0.19.0
WTForms==3.0.1
email-validator==2.0.0